requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.25.0
flask>=2.0.0
streamlit>=1.28.0